LLamaIndex/LangChain/requirements.txt
llama-index==0.10.56
llama-index-llms-langchain==0.3.0
langchain_community==0.2.10
#langchain==0.2.11